Phase B7: extract status handler into src/commands/status.rs
