+++ crates/defra-agent-cli/src/commands/show.rs
@@ -1,9 +1,7 @@
 use anyhow::Result;
 
 use crate::cli::args::{RuntimeShowArgs, ShowCommand};
-use crate::{
-    load_runtime_status_output, print_json, resolve_agent_did, resolve_graphql_endpoint,
-};
+use crate::{print_json, resolve_agent_did, resolve_graphql_endpoint};
 
 pub(crate) async fn dispatch(command: ShowCommand) -> Result<()> {
     match command {
@@ -16,7 +14,9 @@ pub(crate) async fn dispatch(command: ShowCommand) -> Result<()> {
 async fn show_runtime(args: RuntimeShowArgs) -> Result<()> {
     let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
     let agent_did = resolve_agent_did(args.home.as_deref(), args.agent_did.as_deref())?;
-    let output = load_runtime_status_output(args.home.as_deref(), &graphql, &agent_did).await?;
+    let output =
+        crate::commands::status::load_runtime_status_output(args.home.as_deref(), &graphql, &agent_did)
+            .await?;
     print_json(&output)?;
     Ok(())
 }
