+++ crates/defra-agent-cli/src/main.rs
@@ -1,4 +1,4 @@
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeSet;
 use std::fs;
 use std::io::{self, Read};
 use std::net::{IpAddr, Ipv4Addr};
@@ -215,7 +215,7 @@ async fn main() -> Result<()> {
         Command::Chat(args) => commands::chat::chat(args).await,
         Command::P2p { command } => commands::p2p::dispatch(command).await,
         Command::Show { command } => commands::show::dispatch(command).await,
-        Command::Status(args) => status(args).await,
+        Command::Status(args) => commands::status::status(args).await,
         Command::Diagnose(args) => commands::diagnose::diagnose(args).await,
         Command::Config { command } => commands::config::dispatch(command).await,
         Command::Request { command } => commands::request::dispatch(command).await,
@@ -224,199 +224,6 @@ async fn main() -> Result<()> {
     telemetry.shutdown();
     result
 }
-
-async fn status(args: StatusArgs) -> Result<()> {
-    let graphql = resolve_graphql_endpoint(args.graphql.as_deref(), args.home.as_deref())?;
-    let agent_did = resolve_agent_did(args.home.as_deref(), args.agent_did.as_deref())?;
-    let output = load_runtime_status_output(args.home.as_deref(), &graphql, &agent_did).await?;
-    print_json(&output)?;
-    Ok(())
-}
-
-
-pub(crate) async fn load_runtime_status_output(
-    home: Option<&Path>,
-    graphql: &str,
-    agent_did: &str,
-) -> Result<Value> {
-    let unavailable_behaviors = load_live_unavailable_behaviors(graphql, agent_did).await;
-    let query = format!(
-        r#"{{
-            AgentRuntime(
-                filter: {{ agent_did: {{ _eq: "{agent_did}" }} }},
-                limit: 1
-            ) {{
-                agent_did
-                process_state
-                reconcile_phase
-                active_generation
-                router_generation
-                default_behavior_id
-                runnable_behavior_count
-                unavailable_behavior_count
-                last_reconcile_result
-                last_reconcile_error
-                last_reconcile_completed_at
-                updated_at
-            }}
-        }}"#,
-        agent_did = escape_graphql_string(agent_did),
-    );
-    let response = post_graphql(graphql, &query).await?;
-    let runtime_row = response
-        .pointer("/data/AgentRuntime")
-        .and_then(Value::as_array)
-        .and_then(|rows| rows.first())
-        .cloned()
-        .unwrap_or(Value::Null);
-    let home_dir = resolve_home_dir(home);
-    let runtime_state = read_runtime_state(&home_dir)?;
-    let p2p_status = commands::p2p::load_live_http_p2p_status(home, graphql).await;
-    let mut output = json!({
-        "home": home_dir,
-        "graphql": graphql,
-        "agent_did": agent_did,
-        "runtime_state": runtime_state,
-        "runtime": runtime_row,
-        "p2p": p2p_status,
-        "behavior_readiness": if unavailable_behaviors.is_empty() { "ready" } else { "degraded" },
-        "unavailable_behaviors": unavailable_behaviors,
-    });
-    if let Some(map) = output.as_object_mut() {
-        for field in [
-            "process_state",
-            "reconcile_phase",
-            "active_generation",
-            "router_generation",
-            "default_behavior_id",
-            "runnable_behavior_count",
-            "unavailable_behavior_count",
-            "last_reconcile_result",
-            "last_reconcile_error",
-            "last_reconcile_completed_at",
-        ] {
-            map.insert(
-                field.to_string(),
-                runtime_row.get(field).cloned().unwrap_or(Value::Null),
-            );
-        }
-        let p2p_value = map.get("p2p").cloned().unwrap_or(Value::Null);
-        commands::p2p::flatten_p2p_fields(map, &p2p_value);
-    }
-    Ok(output)
-}
-
-async fn load_live_unavailable_behaviors(
-    graphql: &str,
-    agent_did: &str,
-) -> BTreeMap<String, String> {
-    let access = ConfigAccess::Graphql(graphql.to_string());
-    match build_config_export_bundle(&access, agent_did).await {
-        Ok(bundle) => collect_unavailable_behaviors_from_bundle(&bundle),
-        Err(_) => BTreeMap::new(),
-    }
-}
-
-fn collect_unavailable_behaviors_from_bundle(
-    bundle: &ConfigExportBundle,
-) -> BTreeMap<String, String> {
-    let backend_rows = bundle
-        .inference_backends
-        .iter()
-        .filter_map(|row| string_field(row, "backend_id").map(|backend_id| (backend_id, row)))
-        .collect::<BTreeMap<_, _>>();
-    let tool_selection_rows = bundle
-        .tool_selections
-        .iter()
-        .filter_map(|row| string_field(row, "selection_id").map(|selection_id| (selection_id, row)))
-        .collect::<BTreeMap<_, _>>();
-    let inference_profile_rows = bundle
-        .inference_profiles
-        .iter()
-        .filter_map(|row| string_field(row, "profile_id").map(|profile_id| (profile_id, row)))
-        .collect::<BTreeMap<_, _>>();
-
-    let mut unavailable = BTreeMap::new();
-    for behavior in &bundle.agent_behaviors {
-        let Some(behavior_id) = string_field(behavior, "behavior_id") else {
-            continue;
-        };
-        if !bool_field(behavior, "enabled", true) {
-            unavailable.insert(
-                behavior_id.clone(),
-                format!("behavior {behavior_id} is disabled"),
-            );
-            continue;
-        }
-
-        let Some(backend_id) = string_field(behavior, "backend_id") else {
-            unavailable.insert(
-                behavior_id.clone(),
-                format!("behavior {behavior_id} has no backend binding"),
-            );
-            continue;
-        };
-        let Some(backend) = backend_rows.get(&backend_id) else {
-            unavailable.insert(
-                behavior_id.clone(),
-                format!("behavior {behavior_id} references missing backend {backend_id}"),
-            );
-            continue;
-        };
-
-        let probe_status =
-            string_field(backend, "probe_status").unwrap_or_else(|| "unknown".to_string());
-        let backend_enabled = bool_field(backend, "enabled", true);
-        if !backend_enabled || probe_status != "healthy" {
-            unavailable.insert(
-                behavior_id.clone(),
-                format!(
-                    "behavior {behavior_id} backend {backend_id} is unavailable (enabled={backend_enabled} probe_status={probe_status})"
-                ),
-            );
-            continue;
-        }
-
-        if let Some(profile_id) = string_field(behavior, "inference_profile_id") {
-            if !inference_profile_rows.contains_key(&profile_id) {
-                unavailable.insert(
-                    behavior_id.clone(),
-                    format!(
-                        "behavior {behavior_id} references missing inference profile {profile_id}"
-                    ),
-                );
-                continue;
-            }
-        }
-
-        let _tool_selection = match string_field(behavior, "tool_selection_id") {
-            Some(selection_id) => match tool_selection_rows.get(&selection_id) {
-                Some(row) => Some(*row),
-                None => {
-                    unavailable.insert(
-                        behavior_id.clone(),
-                        format!(
-                            "behavior {behavior_id} references missing tool selection {selection_id}"
-                        ),
-                    );
-                    continue;
-                }
-            },
-            None => None,
-        };
-    }
-
-    unavailable
-}
-
-fn string_field(row: &Value, field: &str) -> Option<String> {
-    normalize_optional_string(row.get(field).and_then(Value::as_str))
-}
-
-fn bool_field(row: &Value, field: &str, default: bool) -> bool {
-    row.get(field).and_then(Value::as_bool).unwrap_or(default)
-}
-
 
 pub(crate) fn expand_nonempty_values(values: &[String], flag_name: &str) -> Result<Vec<String>> {
     let values = values
@@ -2209,91 +2016,6 @@ fn server_start_failure_hint(home_dir: &Path) -> String {
 mod tests {
     use super::*;
 
-    fn bundle_with_rows(
-        agent_behaviors: Vec<Value>,
-        tool_selections: Vec<Value>,
-        inference_backends: Vec<Value>,
-        inference_profiles: Vec<Value>,
-    ) -> ConfigExportBundle {
-        ConfigExportBundle {
-            format: CONFIG_EXPORT_FORMAT.to_string(),
-            agent_did: "did:defra-agent:test".to_string(),
-            exported_at: "2026-04-14T00:00:00Z".to_string(),
-            access_mode: "graphql".to_string(),
-            agent_principal: None,
-            agent_behaviors,
-            tool_selections,
-            inference_backends,
-            inference_profiles,
-            tool_service_registries: Vec::new(),
-            scheduled_tasks: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn collect_unavailable_behaviors_from_bundle_reports_config_and_backend_issues() {
-        let bundle = bundle_with_rows(
-            vec![
-                json!({
-                    "behavior_id": "did:defra-agent:test:default",
-                    "enabled": true,
-                    "backend_id": "",
-                    "tool_selection_id": "",
-                    "inference_profile_id": ""
-                }),
-                json!({
-                    "behavior_id": "did:defra-agent:test:ops",
-                    "enabled": true,
-                    "backend_id": "backend-unhealthy",
-                    "tool_selection_id": "",
-                    "inference_profile_id": ""
-                }),
-                json!({
-                    "behavior_id": "did:defra-agent:test:broken-tools",
-                    "enabled": true,
-                    "backend_id": "backend-healthy",
-                    "tool_selection_id": "missing-tools",
-                    "inference_profile_id": ""
-                }),
-            ],
-            Vec::new(),
-            vec![
-                json!({
-                    "backend_id": "backend-unhealthy",
-                    "provider_kind": "OpenAiCompatible",
-                    "enabled": true,
-                    "probe_status": "unknown"
-                }),
-                json!({
-                    "backend_id": "backend-healthy",
-                    "provider_kind": "OpenAiCompatible",
-                    "enabled": true,
-                    "probe_status": "healthy"
-                }),
-            ],
-            Vec::new(),
-        );
-
-        let unavailable = collect_unavailable_behaviors_from_bundle(&bundle);
-        assert_eq!(
-            unavailable.get("did:defra-agent:test:default"),
-            Some(&"behavior did:defra-agent:test:default has no backend binding".to_string())
-        );
-        assert_eq!(
-            unavailable.get("did:defra-agent:test:ops"),
-            Some(
-                &"behavior did:defra-agent:test:ops backend backend-unhealthy is unavailable (enabled=true probe_status=unknown)".to_string()
-            )
-        );
-        assert_eq!(
-            unavailable.get("did:defra-agent:test:broken-tools"),
-            Some(
-                &"behavior did:defra-agent:test:broken-tools references missing tool selection missing-tools".to_string()
-            )
-        );
-    }
-
-    #[test]
     fn sanitize_inference_backend_drops_deprecated_capability_fields() {
         let input = json!({
             "backend_id": "local",
